moved FeaturePackLayoutDescriptionBuilder to FeaturePackLayoutDescription.Builder
